ARC-1135 make JenkinsAppRequest abstract so we can introduce multiple request types
